--- hw11-cache/src/main/java/homework/crm/model/Client.java
@@ -1,46 +1,0 @@
-package ru.otus.necessary.crm.model;
-
-import ru.otus.homework.annotation.Id;
-
-public class Client {
-    @Id
-    private Long id;
-    private String name;
-
-    public Client() {
-    }
-
-    public Client(String name) {
-        this.id = null;
-        this.name = name;
-    }
-
-    public Client(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "Client{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
