--- hw11-cache/src/main/java/homework/crm/model/Manager.java
@@ -1,55 +1,0 @@
-package ru.otus.necessary.crm.model;
-
-import ru.otus.homework.annotation.Id;
-
-public class Manager {
-    @Id
-    private Long no;
-    private String label;
-    private String param1;
-
-    public Manager() {
-    }
-
-    public Manager(String label) {
-        this.label = label;
-    }
-
-    public Manager(Long no, String label, String param1) {
-        this.no = no;
-        this.label = label;
-        this.param1 = param1;
-    }
-
-    public Long getNo() {
-        return no;
-    }
-
-    public void setNo(Long no) {
-        this.no = no;
-    }
-
-    public String getLabel() {
-        return label;
-    }
-
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    public String getParam1() {
-        return param1;
-    }
-
-    public void setParam1(String param1) {
-        this.param1 = param1;
-    }
-
-    @Override
-    public String toString() {
-        return "Manager{" +
-                "no=" + no +
-                ", label='" + label + '\'' +
-                '}';
-    }
-}
