--- hw11-cache/src/main/java/homework/crm/service/DBServiceClient.java
@@ -1,15 +1,0 @@
-package ru.otus.necessary.crm.service;
-
-import ru.otus.necessary.crm.model.Client;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface DBServiceClient {
-
-    Client saveClient(Client client);
-
-    Optional<Client> getClient(long id);
-
-    List<Client> findAll();
-}
