--- hw11-cache/src/main/java/homework/crm/service/DBServiceManager.java
@@ -1,15 +1,0 @@
-package ru.otus.necessary.crm.service;
-
-import ru.otus.necessary.crm.model.Manager;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface DBServiceManager {
-
-    Manager saveManager(Manager client);
-
-    Optional<Manager> getManager(long no);
-
-    List<Manager> findAll();
-}
