--- hw11-cache/src/main/java/homework/crm/service/DbServiceManagerImpl.java
@@ -1,54 +1,0 @@
-package ru.otus.necessary.crm.service;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import ru.otus.necessary.repository.DataTemplate;
-import ru.otus.necessary.sessionmanager.TransactionRunner;
-import ru.otus.necessary.crm.model.Manager;
-import java.util.List;
-import java.util.Optional;
-
-public class DbServiceManagerImpl implements DBServiceManager {
-    private static final Logger log = LoggerFactory.getLogger(DbServiceManagerImpl.class);
-
-    private final DataTemplate<Manager> managerDataTemplate;
-    private final TransactionRunner transactionRunner;
-
-    public DbServiceManagerImpl(TransactionRunner transactionRunner, DataTemplate<Manager> managerDataTemplate) {
-        this.transactionRunner = transactionRunner;
-        this.managerDataTemplate = managerDataTemplate;
-    }
-
-    @Override
-    public Manager saveManager(Manager manager) {
-        return transactionRunner.doInTransaction(connection -> {
-            if (manager.getNo() == null) {
-                var managerNo = managerDataTemplate.insert(connection, manager);
-                var createdManager = new Manager(managerNo, manager.getLabel(), manager.getParam1());
-                log.info("created manager: {}", createdManager);
-                return createdManager;
-            }
-            managerDataTemplate.update(connection, manager);
-            log.info("updated manager: {}", manager);
-            return manager;
-        });
-    }
-
-    @Override
-    public Optional<Manager> getManager(long no) {
-        return transactionRunner.doInTransaction(connection -> {
-            var managerOptional = managerDataTemplate.findById(connection, no);
-            log.info("manager: {}", managerOptional);
-            return managerOptional;
-        });
-    }
-
-    @Override
-    public List<Manager> findAll() {
-        return transactionRunner.doInTransaction(connection -> {
-            var managerList = managerDataTemplate.findAll(connection);
-            log.info("managerList:{}", managerList);
-            return managerList;
-       });
-    }
-}
