--- hw11-cache/src/main/java/homework/demo/DbServiceDemo.java
@@ -1,58 +1,0 @@
-package ru.otus.necessary.demo;
-
-import org.flywaydb.core.Flyway;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import ru.otus.necessary.crm.model.Client;
-import ru.otus.necessary.crm.service.DbServiceClientImpl;
-import ru.otus.necessary.crm.datasource.DriverManagerDataSource;
-import ru.otus.necessary.repository.executor.DbExecutorImpl;
-import ru.otus.necessary.crm.repository.ClientDataTemplateJdbc;
-import ru.otus.necessary.sessionmanager.TransactionRunnerJdbc;
-
-import javax.sql.DataSource;
-
-public class DbServiceDemo {
-    private static final String URL = "jdbc:postgresql://localhost:5430/demoDB";
-    private static final String USER = "usr";
-    private static final String PASSWORD = "pwd";
-
-    private static final Logger log = LoggerFactory.getLogger(DbServiceDemo.class);
-
-    public static void main(String[] args) {
-        var dataSource = new DriverManagerDataSource(URL, USER, PASSWORD);
-        flywayMigrations(dataSource);
-        var transactionRunner = new TransactionRunnerJdbc(dataSource);
-        var dbExecutor = new DbExecutorImpl();
-///
-        var clientTemplate = new ClientDataTemplateJdbc(dbExecutor); //реализация DataTemplate, заточена на Client
-
-///
-        var dbServiceClient = new DbServiceClientImpl(transactionRunner, clientTemplate);
-        dbServiceClient.saveClient(new Client("dbServiceFirst"));
-
-        var clientSecond = dbServiceClient.saveClient(new Client("dbServiceSecond"));
-        var clientSecondSelected = dbServiceClient.getClient(clientSecond.getId())
-                .orElseThrow(() -> new RuntimeException("Client not found, id:" + clientSecond.getId()));
-        log.info("clientSecondSelected:{}", clientSecondSelected);
-///
-        dbServiceClient.saveClient(new Client(clientSecondSelected.getId(), "dbServiceSecondUpdated"));
-        var clientUpdated = dbServiceClient.getClient(clientSecondSelected.getId())
-                .orElseThrow(() -> new RuntimeException("Client not found, id:" + clientSecondSelected.getId()));
-        log.info("clientUpdated:{}", clientUpdated);
-
-        log.info("All clients");
-        dbServiceClient.findAll().forEach(client -> log.info("client:{}", client));
-    }
-
-    private static void flywayMigrations(DataSource dataSource) {
-        log.info("db migration started...");
-        var flyway = Flyway.configure()
-                .dataSource(dataSource)
-                .locations("classpath:/db/migration")
-                .load();
-        flyway.migrate();
-        log.info("db migration finished.");
-        log.info("***");
-    }
-}
