--- hw11-cache/src/main/java/homework/mapper/annotation/Id.java
@@ -1,12 +1,0 @@
-package ru.otus.homework.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target(FIELD)
-@Retention(RUNTIME)
-public @interface Id {
-}
