--- hw11-cache/src/main/java/homework/mapper/entity/sql/EntitySQLMetaData.java
@@ -1,14 +1,0 @@
-package ru.otus.homework.entity.sql;
-
-/**
- * Создает SQL - запросы
- */
-public interface EntitySQLMetaData {
-    String getSelectAllSql();
-
-    String getSelectByIdSql();
-
-    String getInsertSql();
-
-    String getUpdateSql();
-}
