+++ hw11-cache/src/main/java/ru/otus/core/repository/DataTemplateException.java
