+++ hw11-cache/src/main/java/ru/otus/core/sessionmanager/TransactionRunner.java
