+++ hw11-cache/src/main/java/ru/otus/crm/datasource/DriverManagerDataSource.java
@@ -1,0 +1,88 @@
+package ru.otus.crm.datasource;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+
+import javax.sql.DataSource;
+import java.io.PrintWriter;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.logging.Logger;
+
+public class DriverManagerDataSource implements DataSource {
+    private DataSource dataSourcePool;
+
+    public DriverManagerDataSource(String url, String user, String pwd) {
+        createConnectionPool(url, user, pwd);
+    }
+
+    @Override
+    public Connection getConnection() throws SQLException {
+        return dataSourcePool.getConnection();
+    }
+
+    @Override
+    public Connection getConnection(String username, String password) {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public PrintWriter getLogWriter() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public void setLogWriter(PrintWriter out) {
+        throw new UnsupportedOperationException();
+
+    }
+
+    @Override
+    public int getLoginTimeout() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public void setLoginTimeout(int seconds) {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public Logger getParentLogger() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public <T> T unwrap(Class<T> iface) {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public boolean isWrapperFor(Class<?> iface) {
+        throw new UnsupportedOperationException();
+    }
+
+    private void createConnectionPool(String url, String user, String pwd) {
+        var config = new HikariConfig();
+        config.setJdbcUrl(url);
+        //Выставленные значения зависят от контекста, надо подбирать, либо спросить у dba какое максимальное время жизни
+        //сессии к БД проставлено и проставить значения меньше.
+        config.setConnectionTimeout(3000); //ms
+        config.setIdleTimeout(60000); //ms
+        config.setMaxLifetime(600000);//ms
+        config.setAutoCommit(false);
+        config.setMinimumIdle(5);
+        config.setMaximumPoolSize(10);
+        config.setPoolName("DemoHiPool");
+        config.setRegisterMbeans(true);
+
+        config.addDataSourceProperty("cachePrepStmts", "true");
+        config.addDataSourceProperty("prepStmtCacheSize", "250");
+        config.addDataSourceProperty("prepStmtCacheSqlLimit", "2048");
+
+        config.setUsername(user);
+        config.setPassword(pwd);
+
+        dataSourcePool = new HikariDataSource(config);
+    }
+}
