+++ hw11-cache/src/main/java/ru/otus/crm/service/DBServiceClient.java
@@ -1,0 +1,15 @@
+package ru.otus.crm.service;
+
+import ru.otus.crm.model.Client;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface DBServiceClient {
+
+    Client saveClient(Client client);
+
+    Optional<Client> getClient(long id);
+
+    List<Client> findAll();
+}
