+++ hw11-cache/src/main/java/ru/otus/crm/service/DBServiceManager.java
@@ -1,0 +1,15 @@
+package ru.otus.crm.service;
+
+import ru.otus.crm.model.Manager;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface DBServiceManager {
+
+    Manager saveManager(Manager client);
+
+    Optional<Manager> getManager(long no);
+
+    List<Manager> findAll();
+}
