+++ hw11-cache/src/main/java/ru/otus/mapper/annotation/Id.java
@@ -1,0 +1,12 @@
+package ru.otus.mapper.annotation;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target(FIELD)
+@Retention(RUNTIME)
+public @interface Id {
+}
