+++ hw11-cache/src/main/java/ru/otus/mapper/entity/clazz/EntityClassMetaData.java
@@ -1,0 +1,20 @@
+package ru.otus.mapper.entity.clazz;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+import java.util.List;
+
+/**
+ * "Разбирает" объект на составные части
+ */
+public interface EntityClassMetaData<T> {
+    String getName();
+
+    Constructor<T> getConstructor();
+
+    Field getIdField();
+
+    List<Field> getAllFields();
+
+    List<Field> getFieldsWithoutId();
+}
