+++ hw11-cache/src/main/java/ru/otus/mapper/entity/sql/EntitySQLMetaData.java
@@ -1,0 +1,14 @@
+package ru.otus.mapper.entity.sql;
+
+/**
+ * Создает SQL - запросы
+ */
+public interface EntitySQLMetaData {
+    String getSelectAllSql();
+
+    String getSelectByIdSql();
+
+    String getInsertSql();
+
+    String getUpdateSql();
+}
