+++ hw11-cache/src/main/resources/db/migration/V1__initial_schema.sql
@@ -1,0 +1,6 @@
+create table client
+(
+    id   bigserial not null primary key,
+    name varchar(50)
+);
+
