+++ hw11-cache/src/main/resources/docker/runDb.src
@@ -1,0 +1,7 @@
+docker run --rm --name pg-docker \
+-e POSTGRES_PASSWORD=pwd \
+-e POSTGRES_USER=usr \
+-e POSTGRES_DB=demoDB \
+-p 5430:5432 \
+postgres:13
+
